voice_agent/requirements.txt
# Core backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Speech Recognition
openai-whisper>=20231117

# Intent Detection (hybrid)
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
numpy>=1.26.0

# NLP utilities

# Audio processing
pydub>=0.25.1

# Database (business logic)
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# TTS (Coqui - optional, install separately if needed)
# TTS>=0.22.0

# pyttsx3 as lightweight TTS fallback
pyttsx3>=2.90

# Async support
httpx>=0.27.0
aiofiles>=23.2.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
